+++ src/network/bootnodes.go
@@ -1,6 +1,11 @@
 package network
 
-import "encoding/json"
+import (
+	"encoding/json"
+	"fmt"
+)
+
+const BootNodeLocation = "https://raw.githubusercontent.com/lukso-network/network-configs/main/%s/%s/%s"
 
 type Bootnode struct {
 	Consensus string `json:"consensus"`
@@ -8,19 +13,22 @@ type Bootnode struct {
 }
 
 type BootnodeUpdater struct {
-	Chain Chain
+	chainIdentifier string
 }
 
 func NewBootnodeUpdater(chain Chain) BootnodeUpdater {
-	return BootnodeUpdater{chain}
+	return BootnodeUpdater{chain.String()}
+}
+
+func NewBootnodeUpdaterDev(chain Chain, devLocation string) BootnodeUpdater {
+	return BootnodeUpdater{fmt.Sprintf("%s/%s", chain.String(), devLocation)}
 }
 
 func (b BootnodeUpdater) DownloadLatestBootnodes() ([]Bootnode, error) {
-	url, err := getBootnodeUrl(b.Chain)
-	if err != nil {
-		return nil, err
-	}
-	bytes, err := downloadFileOverHttp(url.String())
+	url := fmt.Sprintf(BootNodeLocation, b.chainIdentifier, BootnodesDirectory, BootnodeJSONName)
+
+	fmt.Println("Fetching bootnodes from", url)
+	bytes, err := downloadFileOverHttp(url)
 	if err != nil {
 		return nil, err
 	}