+++ src/network/resource_downloader.go
@@ -11,19 +11,26 @@ const (
 	DockerComposeDefaultName = "docker-compose.yml"
 	DockerGitLocation        = "https://raw.githubusercontent.com/lukso-network/network-configs/main/%s/docker/docker-compose.%s.yml"
 	ConfigGitLocation        = "https://raw.githubusercontent.com/lukso-network/network-configs/main/%s/configs"
+	DevDockerGitLocation     = "https://raw.githubusercontent.com/lukso-network/network-configs/main/dev/%s/docker/docker-compose.%s.yml"
+	DevConfigGitLocation     = "https://raw.githubusercontent.com/lukso-network/network-configs/main/dev/%s/configs"
 )
 
 type ResourceDownloader struct {
 	ConfigLocation string
 	DockerLocation string
-	Chain          Chain
 }
 
 func NewResourceDownloader(chain Chain, clientVersion string) ResourceDownloader {
 	downloader := ResourceDownloader{}
 	downloader.DockerLocation = fmt.Sprintf(DockerGitLocation, chain.String(), clientVersion)
 	downloader.ConfigLocation = fmt.Sprintf(ConfigGitLocation, chain.String())
-	downloader.Chain = chain
+	return downloader
+}
+
+func NewDevResourceDownloader(devLocation, clientVersion string) ResourceDownloader {
+	downloader := ResourceDownloader{}
+	downloader.DockerLocation = fmt.Sprintf(DevDockerGitLocation, devLocation, clientVersion)
+	downloader.ConfigLocation = fmt.Sprintf(DevConfigGitLocation, devLocation)
 	return downloader
 }
 
@@ -56,15 +63,18 @@ func (d ResourceDownloader) downloadDocker() error {
 }
 
 func (d ResourceDownloader) downloadConfig() error {
-	for _, fileName := range ConfigFiles {
-		fileUrl, err := getConfigFileUrl(d.Chain, fileName)
-		if err != nil {
-			return err
-		}
+	for _, fileName := range []string{"config.yaml", "genesis.json", "genesis.ssz"} {
+		fileUrl := fmt.Sprintf("%s/%s", d.ConfigLocation, fileName)
 		fmt.Printf("Fetching %s from %s..... ", fileName, fileUrl)
 		destLocation := path.Join(ConfigDirectory, fileName)
-		if err = downloadFile(fileUrl.String(), destLocation); err != nil {
-			return err
+		if err := downloadFile(fileUrl, destLocation); err != nil {
+			fmt.Printf("unsuccessful\n")
+			// we can ignore the gensis.ssz as a chain can be setup in a way where it is not needed
+			if fileName != "genesis.ssz" {
+				return err
+			} else {
+				continue
+			}
 		}
 		fmt.Printf("success\n")
 	}